- move the email helper to auth.py for simplicity - move the email SMTP server variables to .env for better management of important variables
