+++ app/config.py
@@ -19,10 +19,10 @@ class Config:
     WTF_CSRF_ENABLED = True
 
     # Gmail SMTP server with TLS encryption
-    MAIL_SERVER = "smtp.gmail.com"
-    MAIL_PORT = 587
-    MAIL_USE_TLS = True
-    MAIL_USE_SSL = False
+    MAIL_SERVER = os.environ.get("MAIL_SERVER")
+    MAIL_PORT = int(os.environ.get("MAIL_PORT"))
+    MAIL_USE_TLS = os.environ.get("MAIL_USE_TLS") == "True"
+    MAIL_USE_SSL = os.environ.get("MAIL_USE_SSL") == "True"
     MAIL_USERNAME = os.environ.get("MAIL_USERNAME")
     MAIL_PASSWORD = os.environ.get("MAIL_PASSWORD")
     MAIL_DEFAULT_SENDER = os.environ.get("MAIL_USERNAME")
